refactor : merge 전 center/member 부분 수정 - IsUniqueDto package 이동
