+++ src/main/java/com/prgrms/needit/domain/user/user/dto/IsUniqueRequest.java
@@ -1,4 +1,4 @@
-package com.prgrms.needit.common.domain.dto;
+package com.prgrms.needit.domain.user.user.dto;
 
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
