+++ src/main/java/com/prgrms/needit/domain/user/user/dto/IsUniqueResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.needit.common.domain.dto;
+package com.prgrms.needit.domain.user.user.dto;
 
 public class IsUniqueResponse {
