+++ src/main/java/com/prgrms/needit/domain/user/user/service/UserService.java
@@ -1,8 +1,6 @@
 package com.prgrms.needit.domain.user.user.service;
 
 import com.prgrms.needit.common.config.jwt.JwtTokenProvider;
-import com.prgrms.needit.common.domain.dto.IsUniqueRequest;
-import com.prgrms.needit.common.domain.dto.IsUniqueResponse;
 import com.prgrms.needit.common.enums.UserType;
 import com.prgrms.needit.common.error.ErrorCode;
 import com.prgrms.needit.common.error.exception.NotFoundResourceException;
@@ -11,6 +9,8 @@ import com.prgrms.needit.domain.user.center.repository.CenterRepository;
 import com.prgrms.needit.domain.user.member.entity.Member;
 import com.prgrms.needit.domain.user.member.repository.MemberRepository;
 import com.prgrms.needit.domain.user.user.dto.CurUser;
+import com.prgrms.needit.domain.user.user.dto.IsUniqueRequest;
+import com.prgrms.needit.domain.user.user.dto.IsUniqueResponse;
 import com.prgrms.needit.domain.user.user.dto.LoginRequest;
 import com.prgrms.needit.domain.user.user.dto.TokenResponse;
 import java.util.ArrayList;
